LEP-214 Swap round PropertyCalculator and PropertyAssessor
